fix buddy departure events in macos aim v4.0.0
History of SNAC 0x03,0x0C TLV value:
- Initially empty, which worked for AIM. Discovered that departure
  events in ICQ did not trigger.
- Added a uint8 value (per iserverd), which fixed AIM + ICQ.
- Discovered that departure events didn't work for macOS v4.0.9 client.
  Changing the type from uint8 to uint16 fixed departure events for all
  clients.

+++ foodgroup/buddy.go
@@ -221,7 +221,7 @@ func (s BuddyService) BroadcastBuddyDeparted(ctx context.Context, sess *state.Se
 					TLVList: wire.TLVList{
 						// this TLV needs to be set in order for departure
 						// events to work in ICQ
-						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint8(0)),
+						wire.NewTLV(wire.OServiceUserInfoUserFlags, uint16(0)),
 					},
 				},
 			},
